Refactor segmentation on common method
