+++ src/highdicom/seg/sop.py
@@ -1,6 +1,6 @@
 """Module for SOP classes of the SEG modality."""
+from concurrent.futures import Executor
 import logging
-from concurrent.futures import Executor, Future, ProcessPoolExecutor
 from copy import deepcopy
 from os import PathLike
 import pkgutil
@@ -17,8 +17,6 @@ import numpy as np
 from pydicom.dataelem import DataElement
 from pydicom.dataset import Dataset
 from pydicom.datadict import keyword_for_tag, tag_for_keyword
-from pydicom.encaps import encapsulate, encapsulate_extended
-from pydicom.pixels.utils import pack_bits
 from pydicom.tag import BaseTag, Tag
 from pydicom.uid import (
     ExplicitVRLittleEndian,
@@ -29,7 +27,7 @@ from pydicom.uid import (
     UID,
 )
 from pydicom.sr.codedict import codes
-from pydicom.valuerep import PersonName, format_number_as_ds
+from pydicom.valuerep import PersonName
 from pydicom.sr.coding import Code
 
 from highdicom.image import _Image
@@ -48,7 +46,6 @@ from highdicom.enum import (
     CoordinateSystemNames,
     DimensionOrganizationTypeValues,
 )
-from highdicom.frame import encode_frame
 from highdicom.pr.content import (
     _add_icc_profile_attributes,
     _add_palette_color_lookup_table_attributes,
@@ -66,15 +63,8 @@ from highdicom.seg.enum import (
 from highdicom.seg.utils import iter_segments
 from highdicom.spatial import (
     get_tile_array,
-    get_volume_positions,
-    is_tiled_image,
 )
 from highdicom.sr.coding import CodedConcept
-from highdicom.valuerep import (
-    check_person_name,
-    _check_code_string,
-    _check_long_string,
-)
 from highdicom.volume import (
     ChannelDescriptor,
     Volume,
@@ -84,12 +74,6 @@ from highdicom.volume import (
 
 
 logger = logging.getLogger(__name__)
-
-
-# This code is needed many times in loops so we precompute it
-_DERIVATION_CODE = CodedConcept.from_code(
-    codes.cid7203.SegmentationImageDerivation
-)
 
 
 def _get_unsigned_dtype(max_val: int | np.integer) -> type:
@@ -559,6 +543,9 @@ class Segmentation(_Image):
             pyramid_label=pyramid_label,
             pyramid_uid=pyramid_uid,
         )
+        self.copy_specimen_information(src_img)
+        self.copy_patient_and_study_information(src_img)
+        self._add_contributing_equipment(contributing_equipment, src_img)
 
         # Check segment numbers
         described_segment_numbers = np.array([
@@ -612,29 +599,6 @@ class Segmentation(_Image):
             pixel_array = pixel_array[np.newaxis, ...]
         if pixel_array.ndim not in [3, 4]:
             raise ValueError('Pixel array must be a 2D, 3D, or 4D array.')
-
-        is_tiled = is_tiled_image(src_img)
-        if tile_pixel_array and not is_tiled:
-            raise ValueError(
-                'When argument "tile_pixel_array" is True, the source image '
-                'must be a tiled image.'
-            )
-        if tile_pixel_array and pixel_array.shape[0] != 1:
-            raise ValueError(
-                'When argument "tile_pixel_array" is True, the input pixel '
-                'array must contain only one "frame" representing the '
-                'entire pixel matrix.'
-            )
-
-        # Image Pixel
-        if tile_pixel_array:
-            # By default use the same tile size as the source image (even if
-            # they are not spatially aligned)
-            tile_size = tile_size or (src_img.Rows, src_img.Columns)
-            self.Rows, self.Columns = (tile_size)
-        else:
-            self.Rows = pixel_array.shape[1]
-            self.Columns = pixel_array.shape[2]
 
         # Segmentation Image
         self.ImageType = ['DERIVED', 'PRIMARY']
@@ -703,23 +667,54 @@ class Segmentation(_Image):
             max_described_segment=int(described_segment_numbers.max()),
         )
 
+        self._add_segment_descriptions(
+            segment_descriptions,
+            segmentation_type,
+        )
+
         # Multi-Frame Functional Groups and Multi-Frame Dimensions
         include_segment_number = (
             segmentation_type != SegmentationTypeValues.LABELMAP
         )
+        if include_segment_number:
+            channel_values = described_segment_numbers.tolist()
+
+            # Function to use as callback to add the segment identification
+            def _add_segment_identification(
+                pffg_item: Dataset,
+                segment_number: int,
+            ):
+                identification = Dataset()
+                identification.add(
+                    DataElement(
+                        0x0062000b,  # ReferencedSegmentNumber
+                        'US',
+                        int(segment_number)
+                    )
+                )
+                pffg_item.add(
+                    DataElement(
+                        0x0062000a,  # SegmentIdentificationSequence
+                        'SQ',
+                        [identification]
+                    )
+                )
+                return pffg_item
+
+            add_channel_callback = _add_segment_identification
+        else:
+            channel_values = None
+            add_channel_callback = None
+
+        # TODO factor this into the common method
         self.DimensionIndexSequence = DimensionIndexSequence(
             coordinate_system=self._coordinate_system,
             include_segment_number=include_segment_number,
         )
         dimension_organization = Dataset()
         dimension_organization.DimensionOrganizationUID = \
-            self.DimensionIndexSequence[0].DimensionOrganizationUID
+        self.DimensionIndexSequence[0].DimensionOrganizationUID
         self.DimensionOrganizationSequence = [dimension_organization]
-
-        self._add_segment_descriptions(
-            segment_descriptions,
-            segmentation_type,
-        )
 
         # Checks on pixels and overlap
         pixel_array, segments_overlap = self._check_and_cast_pixel_array(
@@ -730,460 +725,38 @@ class Segmentation(_Image):
         )
         self.SegmentsOverlap = segments_overlap.value
 
-        (
-            plane_positions,
-            plane_orientation,
-            pixel_measures,
-            plane_position_values,
-            plane_sort_index,
-            derivation_source_image_items,
-        ) = self._prepare_spatial_metadata(
-            plane_positions=plane_positions,
-            plane_orientation=plane_orientation,
-            pixel_measures=pixel_measures,
+        def preprocess_channel_callback(
+            plane_array: np.ndarray,
+            segment_number: int,
+            segment_index: int,  # unused
+        ):
+            return self._get_segment_pixel_array(
+                plane_array,
+                segment_number=segment_number,
+                described_segment_numbers=described_segment_numbers,
+                segmentation_type=segmentation_type,
+                max_fractional_value=max_fractional_value,
+                dtype=dtype,
+            )
+
+        self._common_derived_multiframe(
             source_images=source_images,
-            further_source_images=further_source_images or [],
+            pixel_array=pixel_array,
+            dtype=dtype,
+            pixel_measures=pixel_measures,
+            plane_orientation=plane_orientation,
+            plane_positions=plane_positions,
+            omit_empty_frames=omit_empty_frames,
+            workers=workers,
+            dimension_organization_type=dimension_organization_type,
             tile_pixel_array=tile_pixel_array,
             tile_size=tile_size,
-            frame_shape=pixel_array.shape[1:3],
-            number_of_planes=pixel_array.shape[0],
-            dimension_organization_type=dimension_organization_type,
+            further_source_images=further_source_images,
+            use_extended_offset_table=use_extended_offset_table,
+            channel_values=channel_values,
+            add_channel_callback=add_channel_callback,
+            preprocess_channel_callback=preprocess_channel_callback,
         )
-
-        # Shared functional groops
-        sffg_item = Dataset()
-        if (
-            self._coordinate_system is not None and
-            self._coordinate_system == CoordinateSystemNames.PATIENT
-        ):
-            sffg_item.PlaneOrientationSequence = plane_orientation
-
-            # Automatically populate the spacing between slices in the
-            # pixel measures if it was not provided. This is done on the
-            # initial plane positions, before any removals, to give the
-            # receiver more information about how to reconstruct a volume
-            # from the frames in the case that slices are omitted
-            if 'SpacingBetweenSlices' not in pixel_measures[0]:
-                ori = plane_orientation[0].ImageOrientationPatient
-                slice_spacing, _ = get_volume_positions(
-                    image_positions=plane_position_values[:, 0, :],
-                    image_orientation=ori,
-                )
-                if slice_spacing is not None:
-                    pixel_measures[0].SpacingBetweenSlices = (
-                        format_number_as_ds(slice_spacing)
-                    )
-
-        if pixel_measures is not None:
-            sffg_item.PixelMeasuresSequence = pixel_measures
-        self.SharedFunctionalGroupsSequence = [sffg_item]
-
-        # Find indices such that empty planes are removed
-        if omit_empty_frames:
-            if tile_pixel_array:
-                included_plane_indices, is_empty = \
-                    self._get_nonempty_tile_indices(
-                        pixel_array,
-                        plane_positions=plane_positions,
-                        rows=self.Rows,
-                        columns=self.Columns,
-                    )
-            else:
-                included_plane_indices, is_empty = \
-                    self._get_nonempty_plane_indices(pixel_array)
-            if is_empty:
-                # Cannot omit empty frames when all frames are empty
-                omit_empty_frames = False
-                included_plane_indices = range(len(plane_positions))
-            else:
-                # Remove all empty plane positions from the list of sorted
-                # plane position indices
-                included_plane_indices_set = set(included_plane_indices)
-                plane_sort_index = [
-                    ind for ind in plane_sort_index
-                    if ind in included_plane_indices_set
-                ]
-        else:
-            included_plane_indices = range(len(plane_positions))
-
-        # Dimension Organization Type
-        dimension_organization_type = self._check_tiled_dimension_organization(
-            dimension_organization_type=dimension_organization_type,
-            is_tiled=is_tiled,
-            omit_empty_frames=omit_empty_frames,
-            plane_positions=plane_positions,
-            tile_pixel_array=tile_pixel_array,
-            rows=self.Rows,
-            columns=self.Columns,
-        )
-
-        if (
-            self._coordinate_system is not None and
-            dimension_organization_type !=
-            DimensionOrganizationTypeValues.TILED_FULL
-        ):
-            # Get unique values of attributes in the Plane Position Sequence or
-            # Plane Position Slide Sequence, which define the position of the
-            # plane with respect to the three dimensional patient or slide
-            # coordinate system, respectively. These can subsequently be used
-            # to look up the relative position of a plane relative to the
-            # indexed dimension.
-            unique_dimension_values = [
-                np.unique(
-                    plane_position_values[included_plane_indices, index],
-                    axis=0
-                )
-                for index in range(plane_position_values.shape[1])
-            ]
-        else:
-            unique_dimension_values = [None]
-
-        if self._coordinate_system == CoordinateSystemNames.PATIENT:
-            inferred_dim_org_type = None
-
-            # To be considered "3D", a segmentation should have frames that are
-            # differentiated only by location. This rules out any non-labelmap
-            # segmentations with more than a single segment.
-            # Further, only segmentation with multiple spatial positions in the
-            # final segmentation should be considered to have 3D dimension
-            # organization type
-            if (
-                len(included_plane_indices) > 1 and
-                (
-                    segmentation_type == SegmentationTypeValues.LABELMAP or
-                    len(described_segment_numbers) == 1
-                )
-            ):
-                # Calculate the spacing using only the included planes, and
-                # enforce ordering
-                ori = plane_orientation[0].ImageOrientationPatient
-                spacing, _ = get_volume_positions(
-                    image_positions=plane_position_values[
-                        included_plane_indices, 0, :
-                    ],
-                    image_orientation=ori,
-                    sort=False,
-                )
-                if spacing is not None and spacing > 0.0:
-                    inferred_dim_org_type = (
-                        DimensionOrganizationTypeValues.THREE_DIMENSIONAL
-                    )
-
-            if (
-                dimension_organization_type ==
-                DimensionOrganizationTypeValues.THREE_DIMENSIONAL
-            ) and inferred_dim_org_type is None:
-                raise ValueError(
-                    'Dimension organization "3D" has been specified, '
-                    'but the source image is not a regularly-spaced 3D '
-                    'volume.'
-                )
-            dimension_organization_type = inferred_dim_org_type
-
-        if dimension_organization_type is not None:
-            self.DimensionOrganizationType = dimension_organization_type.value
-
-        if (
-            self._coordinate_system is not None and
-            self._coordinate_system == CoordinateSystemNames.SLIDE
-        ):
-            plane_position_names = (
-                self.DimensionIndexSequence.get_index_keywords()
-            )
-            row_dim_index = plane_position_names.index(
-                'RowPositionInTotalImagePixelMatrix'
-            )
-            col_dim_index = plane_position_names.index(
-                'ColumnPositionInTotalImagePixelMatrix'
-            )
-
-        is_encaps = self.file_meta.TransferSyntaxUID.is_encapsulated
-        process_pool: Executor | None = None
-
-        if not isinstance(workers, (int, Executor)):
-            raise TypeError(
-                'Argument "workers" must be of type int or '
-                'concurrent.futures.Executor (or a derived class).'
-            )
-        using_multiprocessing = (
-            isinstance(workers, Executor) or workers != 0
-        )
-
-        # List of frames. In the case of native transfer syntaxes, we will
-        # collect a list of frames as flattened NumPy arrays for bitpacking at
-        # the end. In the case of encapsulated transfer syntaxes with no
-        # workers, we will accumulate a list of encoded frames to encapsulate
-        # at the end
-        frames: list[bytes] | list[np.ndarray] = []
-
-        # In the case of native encoding when the number pixels in a frame is
-        # not a multiple of 8. This array carries "leftover" pixels that
-        # couldn't be encoded in previous iterations, to future iterations. This
-        # saves having to keep the entire un-endoded array in memory, which can
-        # get extremely heavy on memory in the case of very large arrays
-        remainder_pixels = np.empty((0, ), dtype=np.uint8)
-
-        if is_encaps:
-            if using_multiprocessing:
-                # In the case of encapsulated transfer syntaxes with multiple
-                # workers, we will accumulate a list of encoded frames to
-                # encapsulate at the end
-                frame_futures: list[Future] = []
-
-                # Use the existing executor or create one
-                if isinstance(workers, Executor):
-                    process_pool = workers
-                else:
-                    # If workers is negative, pass None to use all processors
-                    process_pool = ProcessPoolExecutor(
-                        workers if workers > 0 else None
-                    )
-
-            # Parameters to use when calling the encode_frame function in
-            # either of the above two cases
-            encode_frame_kwargs = dict(
-                transfer_syntax_uid=self.file_meta.TransferSyntaxUID,
-                bits_allocated=self.BitsAllocated,
-                bits_stored=self.BitsStored,
-                photometric_interpretation=self.PhotometricInterpretation,
-                pixel_representation=self.PixelRepresentation
-            )
-        else:
-            if using_multiprocessing:
-                warnings.warn(
-                    "Setting workers != 0 or passing an instance of "
-                    "concurrent.futures.Executor when using a non-encapsulated "
-                    "transfer syntax has no effect.",
-                    UserWarning,
-                    stacklevel=2,
-                )
-                using_multiprocessing = False
-
-        # Information about individual frames is placed into the
-        # PerFrameFunctionalGroupsSequence. Note that a *very* significant
-        # efficiency gain is observed when building this as a Python list
-        # rather than a pydicom sequence, and then converting to a pydicom
-        # sequence at the end
-        pffg_sequence: list[Dataset] = []
-
-        # We want the larger loop to work in the labelmap cases (where segments
-        # are dealt with together) and the other cases (where segments are
-        # dealt with separately). So we define a suitable iterable here for
-        # each case
-        segments_iterable = (
-            [None] if segmentation_type == SegmentationTypeValues.LABELMAP
-            else described_segment_numbers
-        )
-
-        for segment_number in segments_iterable:
-
-            for plane_dim_ind, plane_index in enumerate(plane_sort_index, 1):
-
-                if tile_pixel_array:
-                    if (
-                        dimension_organization_type ==
-                        DimensionOrganizationTypeValues.TILED_FULL
-                    ):
-                        row_offset = int(
-                            plane_position_values[plane_index, row_dim_index]
-                        )
-                        column_offset = int(
-                            plane_position_values[plane_index, col_dim_index]
-                        )
-                    else:
-                        pos = plane_positions[plane_index][0]
-                        row_offset = pos.RowPositionInTotalImagePixelMatrix
-                        column_offset = (
-                            pos.ColumnPositionInTotalImagePixelMatrix
-                        )
-
-                    plane_array = get_tile_array(
-                        pixel_array[0],
-                        row_offset=row_offset,
-                        column_offset=column_offset,
-                        tile_rows=self.Rows,
-                        tile_columns=self.Columns,
-                    )
-                else:
-                    # Select the relevant existing frame
-                    plane_array = pixel_array[plane_index]
-
-                if segment_number is None:
-                    # Deal with all segments at once
-                    segment_array = plane_array
-                else:
-                    # Pixel array for just this segment and this position
-                    segment_array = self._get_segment_pixel_array(
-                        plane_array,
-                        segment_number=segment_number,
-                        described_segment_numbers=described_segment_numbers,
-                        segmentation_type=segmentation_type,
-                        max_fractional_value=max_fractional_value,
-                        dtype=dtype,
-                    )
-
-                # Even though completely empty planes were removed earlier,
-                # there may still be planes in which this specific segment is
-                # absent. Such frames should be removed
-                if segment_number is not None:
-                    if omit_empty_frames and not np.any(segment_array):
-                        logger.debug(
-                            f'skip empty plane {plane_index} of segment '
-                            f'#{segment_number}'
-                        )
-                        continue
-
-                # Log a debug message
-                if segment_number is None:
-                    msg = f'add plane #{plane_index}'
-                else:
-                    msg = (
-                        f'add plane #{plane_index} for segment '
-                        f'#{segment_number}'
-                    )
-                logger.debug(msg)
-
-                if (
-                    dimension_organization_type !=
-                    DimensionOrganizationTypeValues.TILED_FULL
-                ):
-                    # No per-frame functional group for TILED FULL
-
-                    # Get the item of the PerFrameFunctionalGroupsSequence for
-                    # this segmentation frame
-                    if self._coordinate_system is not None:
-                        plane_pos_val = plane_position_values[plane_index]
-                        if (
-                            self._coordinate_system ==
-                            CoordinateSystemNames.SLIDE
-                        ):
-                            try:
-                                dimension_index_values = [
-                                    int(
-                                        np.where(
-                                            unique_dimension_values[idx] == pos
-                                        )[0][0] + 1
-                                    )
-                                    for idx, pos in enumerate(plane_pos_val)
-                                ]
-                            except IndexError as error:
-                                raise IndexError(
-                                    'Could not determine position of plane '
-                                    f'#{plane_index} in three dimensional '
-                                    'coordinate system based on dimension '
-                                    f'index values: {error}'
-                                ) from error
-                        else:
-                            dimension_index_values = [plane_dim_ind]
-                    else:
-                        if segmentation_type == SegmentationTypeValues.LABELMAP:
-                            # Here we have to use the "Frame Label" dimension
-                            # value (which is used just to have one index since
-                            # Referenced Segment cannot be used)
-                            dimension_index_values = [1]
-                        else:
-                            dimension_index_values = []
-
-                    plane_derivation_items = (
-                        derivation_source_image_items[plane_index]
-                        if derivation_source_image_items is not None else None
-                    )
-                    pffg_item = self._get_pffg_item(
-                        segment_number=segment_number,
-                        dimension_index_values=dimension_index_values,
-                        plane_position=plane_positions[plane_index],
-                        derivation_source_image_items=plane_derivation_items,
-                        coordinate_system=self._coordinate_system,
-                    )
-                    pffg_sequence.append(pffg_item)
-
-                # Add the segmentation pixel array for this frame to the list
-                if is_encaps:
-                    if process_pool is None:
-                        # Encode this frame and add resulting bytes to the list
-                        # for encapsulation at the end
-                        frames.append(
-                            encode_frame(
-                                segment_array,
-                                **encode_frame_kwargs,
-                            )
-                        )
-                    else:
-                        # Submit this frame for encoding this frame and add the
-                        # future to the list for encapsulation at the end
-                        future = process_pool.submit(
-                            encode_frame,
-                            array=segment_array,
-                            **encode_frame_kwargs,
-                        )
-                        frame_futures.append(future)
-                else:
-                    flat_array = segment_array.flatten()
-                    if (
-                        self.SegmentationType ==
-                        SegmentationTypeValues.BINARY.value and
-                        (self.Rows * self.Columns) // 8 != 0
-                    ):
-                        # Need to encode a multiple of 8 pixels at a time
-                        full_array = np.concatenate(
-                            [remainder_pixels, flat_array]
-                        )
-                        # Round down to closest multiple of 8
-                        n_pixels_to_take = 8 * (len(full_array) // 8)
-                        to_encode = full_array[:n_pixels_to_take]
-                        remainder_pixels = full_array[n_pixels_to_take:]
-                    else:
-                        # Simple - each frame can be individually encoded
-                        to_encode = flat_array
-
-                    frames.append(self._encode_pixels_native(to_encode))
-
-        if (
-            dimension_organization_type !=
-            DimensionOrganizationTypeValues.TILED_FULL
-        ):
-            self.PerFrameFunctionalGroupsSequence = pffg_sequence
-
-        if is_encaps:
-            if process_pool is not None:
-                frames = [
-                    fut.result() for fut in frame_futures
-                ]
-
-                # Shutdown the pool if we created it, otherwise it is the
-                # caller's responsibility
-                if process_pool is not workers:
-                    process_pool.shutdown()
-
-            # Encapsulate all pre-compressed frames
-            self.NumberOfFrames = len(frames)
-            if use_extended_offset_table:
-                (
-                    self.PixelData,
-                    self.ExtendedOffsetTable,
-                    self.ExtendedOffsetTableLengths,
-                ) = encapsulate_extended(frames)
-            else:
-                self.PixelData = encapsulate(frames)
-        else:
-            self.NumberOfFrames = len(frames)
-
-            # May need to add in a final set of pixels
-            if len(remainder_pixels) > 0:
-                frames.append(self._encode_pixels_native(remainder_pixels))
-
-            self.PixelData = b''.join(frames)
-
-        # Add a null trailing byte if required
-        if len(self.PixelData) % 2 == 1:
-            self.PixelData += b'0'
-
-        self.copy_specimen_information(src_img)
-        self.copy_patient_and_study_information(src_img)
-        self._add_contributing_equipment(contributing_equipment, src_img)
-
-        # Build lookup tables for efficient decoding
-        self._build_luts()
 
     def add_segments(
         self,
@@ -1839,178 +1412,6 @@ class Segmentation(_Image):
                     segment_array *= int(max_fractional_value)
 
         return segment_array
-
-    @staticmethod
-    def _get_pffg_item(
-        segment_number: int | None,
-        dimension_index_values: list[int],
-        plane_position: PlanePositionSequence,
-        derivation_source_image_items: Sequence[Dataset] | None,
-        coordinate_system: CoordinateSystemNames | None,
-    ) -> Dataset:
-        """Get a single item of the Per Frame Functional Groups Sequence.
-
-        This is a helper method used in the constructor.
-
-        Parameters
-        ----------
-        segment_number: Optional[int]
-            Segment number of this segmentation frame. If None, this is a
-            LABELMAP segmentation in which each frame has no segment number.
-        dimension_index_values: List[int]
-            Dimension index values (except segment number) for this frame.
-        plane_position: highdicom.seg.PlanePositionSequence
-            Plane position of this frame.
-        derivation_source_image_items: Sequence[pydicom.Dataset] | None
-            Items of the Source Image Sequence, to place into the Derivation
-            Image Sequence, if any, for this frame.
-        coordinate_system: Optional[highdicom.CoordinateSystemNames]
-            Coordinate system used, if any.
-
-        Returns
-        -------
-        pydicom.Dataset
-            Dataset representing the item of the Per Frame Functional Groups
-            Sequence for this segmentation frame.
-
-        """
-        # NB this function is called many times in a loop when there are a
-        # large number of frames, and has been observed to dominate the
-        # creation time of some segmentations. Therefore we use low-level
-        # pydicom primitives to improve performance as much as possible
-        pffg_item = Dataset()
-        frame_content_item = Dataset()
-
-        if segment_number is None:
-            all_index_values = dimension_index_values
-        else:
-            all_index_values = [int(segment_number)] + dimension_index_values
-
-        frame_content_item.add(
-            DataElement(
-                0x00209157,  # DimensionIndexValues
-                'UL',
-                all_index_values,
-            )
-        )
-
-        if segment_number is None and coordinate_system is None:
-            # If this is an labelmap segmentation of an image that has no frame
-            # of reference, we need to create a dummy frame label to be pointed
-            # to as a dimension index because there is nothing else appropriate
-            # to use
-            frame_content_item.add(
-                DataElement(
-                    0x00209453,  # FrameLabel
-                    'LO',
-                    "Segmentation Frame",
-                )
-            )
-
-        pffg_item.add(
-            DataElement(
-                0x00209111,  # FrameContentSequence
-                'SQ',
-                [frame_content_item]
-            )
-        )
-
-        if coordinate_system is not None:
-            if coordinate_system == CoordinateSystemNames.SLIDE:
-                pffg_item.add(
-                    DataElement(
-                        0x0048021a,  # PlanePositionSlideSequence
-                        'SQ',
-                        plane_position
-                    )
-                )
-            else:
-                pffg_item.add(
-                    DataElement(
-                        0x00209113,  # PlanePositionSequence
-                        'SQ',
-                        plane_position
-                    )
-                )
-
-        if (
-            derivation_source_image_items is not None and
-            len(derivation_source_image_items) > 0
-        ):
-            derivation_image_item = Dataset()
-            derivation_image_item.add(
-                DataElement(
-                    0x00089215,  # DerivationCodeSequence
-                    'SQ',
-                    [_DERIVATION_CODE]
-                )
-            )
-
-            derivation_image_item.add(
-                DataElement(
-                    0x00082112,  # SourceImageSequence
-                    'SQ',
-                    derivation_source_image_items,
-                )
-            )
-            pffg_item.add(
-                DataElement(
-                    0x00089124,  # DerivationImageSequence
-                    'SQ',
-                    [derivation_image_item]
-                )
-            )
-        else:
-            # Determining the source images that map to the frame is not
-            # always trivial. Since DerivationImageSequence is a type 2
-            # attribute, we leave its value empty.
-            pffg_item.add(
-                DataElement(
-                    0x00089124,  # DerivationImageSequence
-                    'SQ',
-                    []
-                )
-            )
-
-        if segment_number is not None:
-            identification = Dataset()
-            identification.add(
-                DataElement(
-                    0x0062000b,  # ReferencedSegmentNumber
-                    'US',
-                    int(segment_number)
-                )
-            )
-            pffg_item.add(
-                DataElement(
-                    0x0062000a,  # SegmentIdentificationSequence
-                    'SQ',
-                    [identification]
-                )
-            )
-
-        return pffg_item
-
-    def _encode_pixels_native(self, planes: np.ndarray) -> bytes:
-        """Encode pixel planes using a native transfer syntax.
-
-        Parameters
-        ----------
-        planes: numpy.ndarray
-            Array representing one or more segmentation image planes. If
-            multiple image planes, planes stacked down the first dimension
-            (index 0).
-
-        Returns
-        -------
-        bytes
-            Encoded pixels
-
-        """
-        if self.SegmentationType == SegmentationTypeValues.BINARY.value:
-            return pack_bits(planes, pad=False)
-        else:
-            return planes.tobytes()
 
     @classmethod
     def from_dataset(
